+++ RELEASENOTES.md
@@ -1,5 +1,10 @@
 # Release notes
 
+## v2.4.1
+ - Fix detection of rebuffering after seeking
+ - Use a random UUID stored in shared preferences for mux_viewer_id
+ - Fix sending of `view_session_id`
+
 ## v2.4.0
  - Fix an issue where additional icons and image files were included
  - Fix an issue where the application would crash on Android 11